Broke down the monolithic script file into several.
