+++ MeetingLocation.Go/DbSettings.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"database/sql"
+)
+
+var db *sql.DB
+
+var server = "."
+var port = 1433
+var user = "saNonWindows"
+var password = "yq$xiElqo988NCV5IsTY"
+var database = "MeetingLocation"
